visualizer_python/requirements.txt
pyserial~=3.5
termcolor~=2.2
